# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vive Hidalgo API",
        "description": "API para la plataforma de turismo Vive Hidalgo",
        "contact": {
            "name": "Soporte Vive Hidalgo",
            "email": "admin@vivehidalgo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Main API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/audit/logs": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Obtener logs de auditoría",
                "description": "Retorna los logs de auditoría con filtros y paginación",
                "operationId": "getAuditLogs",
                "parameters": [
                    {
                        "name": "event_type",
                        "in": "query",
                        "description": "Filtrar por tipo de evento",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filtrar por usuario",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "auditable_type",
                        "in": "query",
                        "description": "Filtrar por tipo de modelo auditado",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Fecha de inicio (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Fecha de fin (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logs obtenidos exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logs obtenidos exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "logs": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/AuditLog"
                                                    }
                                                },
                                                "pagination": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/audit/stats": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Obtener estadísticas de auditoría",
                "description": "Retorna estadísticas y métricas de auditoría",
                "operationId": "getAuditStats",
                "parameters": [
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Número de días para las estadísticas",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estadísticas obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Estadísticas obtenidas exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_logs": {
                                                    "type": "integer"
                                                },
                                                "events_by_type": {
                                                    "type": "object"
                                                },
                                                "top_users": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "user_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "user_name": {
                                                                "type": "string",
                                                                "example": "Juan Pérez"
                                                            },
                                                            "activity_count": {
                                                                "type": "integer",
                                                                "example": 5
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "recent_activity": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "event_type": {
                                                                "type": "string",
                                                                "example": "created"
                                                            },
                                                            "event_description": {
                                                                "type": "string",
                                                                "example": "Creó"
                                                            },
                                                            "user_name": {
                                                                "type": "string",
                                                                "example": "Juan Pérez"
                                                            },
                                                            "auditable_name": {
                                                                "type": "string",
                                                                "example": "Hotel Real"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Creó un destino"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-06-24 12:00:00"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/audit/logs/{id}": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Obtener log de auditoría específico",
                "description": "Retorna un log de auditoría específico con todos sus detalles",
                "operationId": "getAuditLog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del log de auditoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Log obtenido exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AuditLog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Log no encontrado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/audit/logs/clean": {
            "delete": {
                "tags": [
                    "Audit"
                ],
                "summary": "Limpiar logs antiguos",
                "description": "Elimina logs de auditoría más antiguos que el número de días especificado",
                "operationId": "cleanAuditLogs",
                "parameters": [
                    {
                        "name": "days_to_keep",
                        "in": "query",
                        "description": "Número de días a mantener",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 90
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logs limpiados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logs limpiados exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "deleted_count": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuditLog": {
                "title": "Audit Log",
                "required": [
                    "id",
                    "event_type",
                    "user_id",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "event_type": {
                        "type": "string",
                        "example": "created"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "auditable_type": {
                        "type": "string",
                        "example": "App\\\\Models\\\\User"
                    },
                    "auditable_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "type": "string",
                        "example": "Creó un destino"
                    },
                    "ip_address": {
                        "type": "string",
                        "example": "127.0.0.1"
                    },
                    "user_agent": {
                        "type": "string",
                        "example": "Mozilla/5.0"
                    },
                    "url": {
                        "type": "string",
                        "example": "/api/v1/destinos"
                    },
                    "method": {
                        "type": "string",
                        "example": "POST"
                    },
                    "old_values": {
                        "type": "object"
                    },
                    "new_values": {
                        "type": "object"
                    },
                    "metadata": {
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-24T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-24T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}